Move tint selector to shared sibling directory
