+++ docs/plans/2026-05-16-block-generator-updates-v2-session-handover.md
@@ -5,14 +5,17 @@
 - The sibling project lives at `/Users/kevanstannard/dev/pixel-papercraft-generators-ninjolas`.
 - This branch is migrating block-generator and texture-picker features from that sibling project into the current repo.
 - The sibling implementation has drifted and accumulated enough rough edges that it is no longer a good base for continued work, so the features are being rewritten here instead of extended in place.
-- The block generator update work is committed on `block-generator-updates-v2` at `64876d5`.
+- The block generator update work is committed on `block-generator-updates-v2` and now includes the picker/texture-data boundary moves made in this repo.
 - The remaining untracked docs were cleaned up so this is the only handover document left.
 - The current codebase already includes:
   - shared picker flip support
   - `SelectedTexture.flip`
   - the shared `Icon` button size
   - the shared generator texture picker primitives now live in `src/generators/_common/texturePicker`
+  - the shared tint selector primitive now lives in `src/generators/_common/tintSelector`
   - shared texture data and frame-label logic now live in `src/generators/_common/textureData`
+  - the builder directory no longer references anything under `src/generators/_common`
+  - `TexturePicker` and `TintSelector` are now separate shared primitives
   - `SelectedTextureWithBlend` plumbing for the block flow
   - block face rendering that combines rotation, flip, and blend
   - renderer-level and Playwright visual regression coverage for the block generator
@@ -24,8 +27,11 @@
 
 - Picker controls now support rotation, horizontal flip, vertical flip, and erase/reset.
 - The shared texture picker implementation has moved out of `src/builder/ui` and into `src/generators/_common/texturePicker`.
+- The shared tint selector implementation has moved out of `src/generators/minecraftBlock` and into `src/generators/_common/tintSelector`.
 - Shared texture frame data moved out of `src/builder/modules` and into `src/generators/_common/textureData`.
 - The block generator stores the selected texture plus blend/tint state.
+- The block and item texture version files now resolve their shared texture data and texture assets without depending on `src/builder/ui`.
+- The block picker composes `TexturePicker` and `TintSelector` directly; the item picker still omits tint UI.
 - Block face rendering forwards the full orientation state to the renderer.
 - The block texture picker preview now shows the selected tint color.
 - Texture picker search now normalizes underscores to spaces, so `grass block top` matches `grass_block_top`.
@@ -59,9 +65,9 @@
    - Prefer a small unit test, and only add more Playwright coverage if the visual output actually changes.
 
 3. Review the remaining migration items from the sibling project one at a time.
-   - tint selector and the rest of the shared picker composition
+   - keep `TexturePicker` and `TintSelector` as sibling primitives
+   - wire item tint support through the shared tint selector only if the item flow needs it
    - shared texture version assembly
-   - item generator tint support through the shared texture picker
 
 ## Verification Notes
 
@@ -69,6 +75,9 @@
 - `npm run lint`
 - `npx vitest run src/generators/_common/textureData.test.ts`
 - `npx vitest run src/generators/_common/texturePicker/textureSearch.test.ts src/generators/_common/texturePicker/flip.test.ts src/generators/_common/texturePicker/texturePicker.test.ts`
+- `npx vitest run src/generators/_common/tintSelector/tintSelector.test.ts`
+- `npx vitest run src/generators/minecraftBlock/textureVersions.test.ts`
+- `npx vitest run src/generators/minecraftBlock/textureVersions.test.ts src/generators/minecraftItem/ui/textureVersions.test.ts`
 - `npx vitest run src/builder/modules/renderers/drawTexture.test.ts src/generators/minecraftBlock/face.test.ts src/generators/_common/texturePicker/flip.test.ts src/builder/ui/button/buttonStyles.test.ts`
 - `npx playwright test tests/generators/minecraftBlockGenerator/minecraftBlockGenerator.spec.ts`
 - `npx playwright test tests/generators/minecraftBlockGenerator/minecraftBlockGenerator.spec.ts -g "shows the before and after tinting on the page"`
@@ -79,4 +88,24 @@
 - Inspect the existing changes before adding new ones.
 - Use the existing screenshot harness rather than introducing a new visual testing system.
 - Use the `testing` generator as the first home for shared visual regressions.
-- The next small task is moving the tint selector and shared picker composition into `_common`.
+- The next small task is keeping the tint-selector path separate from the texture-picker path while deciding whether item tint support is actually needed.
+
+## Next Session Checklist
+
+1. Read these files first:
+   - `src/generators/minecraftBlock/texturePicker.tsx`
+   - `src/generators/minecraftItem/ui/texturePicker.tsx`
+   - `src/generators/_common/texturePicker/texturePicker.tsx`
+   - `src/generators/_common/tintSelector/tintSelector.tsx`
+
+2. Keep `TexturePicker` and `TintSelector` as separate shared primitives.
+
+3. Add item tint support only if the item flow genuinely needs it.
+
+4. Add or update focused tests before changing behavior:
+   - a unit test for the shared tint component, if it changes
+   - a small picker test if the shared picker API changes
+
+5. Verify the narrowest relevant surface first:
+   - `npx vitest run src/generators/_common/tintSelector/*.test.ts src/generators/_common/texturePicker/*.test.ts`
+   - then the block/item texture-version tests if imports change
