+++ src/generators/_common/tintSelector/tints.ts
